requirements.txt
flask
pandas
sentence-transformers
scikit-learn
google-generativeai
numpy
os-sys
regex
gunicorn  # For deployment (optional)
